KTO-1037: Siirretty KoulutuksenAlkamiskausi case class domain/package.scala-tiedostoon, jotta sitä voidaan käyttää myös toteutuksella ja hakukohteella.
